engine/requirements.txt
asyncio
aiohttp
pydantic
python-dotenv
colorama
cryptography
groq
google-genai
supabase
psutil
ddgs>=4.1.0  # DuckDuckGo search (replaces duckduckgo-search)
numpy
ruff
black
rich>=13.0.0  # Modern CLI display system
